Improvements to `NoMoreThanKInARow` constraint
1. I changed the constraint in `example_programs/simple_stroop.py`
because I don't think specifying an individual level is sufficient to
uniquely identify a given level within an experiment.

For example, giving "red" as the level isn't specific enough, as that
could be referring to the "red" level in the `color` or `text`
factors. If constraining just a specific level is desired, then I
think we need to require a tuple of `(<factor name>, <level name>)`.

2. I made `desugar` smart enough to deal with both individual levels,
as well as entire factors when desugaring `NoMoreThanKInARow`
constraints. When an entire factor is given, `desugar` will expand
that into each `(<factor name>, <level name>)` tuple, and generate the
constraints from there. (I believe that's what is desired in some of
the examples, like [this one][1])

[1]: https://github.com/anniecherk/sweetpea-py/blob/002773e3e04d97ba14f1393a48cadd4bc34a7987/example_programs/task_switching_simple.py#L112-L113