Review use of `fresh` variable
Based on my understanding of the functions (which may, of course, be
flawed), the only functions that needs to mutate the fresh variable is
`desugar_full_crossing`, so I removed `fresh` from the signatures of
the other functions that were using it. (`desugar_derivation`,
`desugar_consistency`, `desguar_nomorethankinarow`)

Additionally, the update to `fresh` inside `desugar_full_crossing`
wasn't being returned to the caller, so it was being lost. I repaired
that by adding it to the tuple that was returned.