Invoke unigen and extract solutions
This adds the next step to `synthesize_trials`, which is to feed the
generated CNF file to `unigen`, and then parse the satisfying
assignments that it produces.

The last step will be to implement `decode`, which will map the
satisfying assignment back into the original language of the
experiment specification.