Conform to new JSON contract
We've made the backend server expect a `cnfs` key in the JSON object
which contains the CNF clause generated by the frontend.

As a result, the frontend no longer needs to turn CNFs into DIMACs, but
it _does_ need to include them in the JSON. For example:

```
(1 || 2 || 3) && (4 || 5 || 6)
```

should become a list of lists:

```
[
  [1, 2, 3],
  [4, 5, 6]
]
```

This makes the necessary changes to the frontend. `synthesize_trials`
will be broken until we make it start relying on the backend server.