Restructure for distribution on pypi.org
This reshapes the file organization and adds the neccessary files so
that `sweetpea` can be distributed via PyPi. (`pip install sweetpea`)

It takes a couple steps to upload a new release, so I'm still planning
on adding a script to help with that.

Also, the final package isn't working yet, hopefully I'll be able to fix
it tonight.