Integrate new logic module
This replaces the current CNF functionality with usage of the logic
module. The primary benefit is that the CNF conversion in the logic
module avoids an exponential explosion in the length of the fomrula, at
the cost of adding additional variables.

This should significantly improve the performance of experiments with
more than just a handful of variables.