Reduce number of samples (for now)
The new logic module is generating formulas more quickly, (O(n^3),
instead of O(n^n)) but `unigen` appears to be taking a _long_ time to
sample the longer formulas. Reducing the sample count should help with
this, at least while we're tuning.

Long term, we should really figure out a better way to configure the
unigen options. Environment variables are the obvious solution, but I
wonder if our users would like that.