Package CNF conversion function with block
In order to get better data/understanding of how the CNF conversion
choice impacts the number of solutions to the formula and conversion
time, I need to be able to specify the CNF conversion function that I
want to be used.

This adds a field to the `HLBlock` named tuple to hold this function. By
default, it will use the only function available at the moment:
`to_cnf_switching`. I anticipate the addition of two other functions:
`to_cnf_naive`, and `to_cnf_tseitin`.

I'm not sure this is the best place for it, but it works well for what
we've done so far.