Add function for non-uniform sampling
This adds a new function for non-uniform sampling:
`synthesize_trials_non_uniform`. This function takes a block and the
desired number of trials, and then synthesizes that number of trials by
repeatedly invoking a SAT solver and then updating the formula to
disallow previously discovered solutions.

In this way, users can start generating trials and getting faster
feedback on their designs, even though the trials may not be uniformly
sampled from the solution space.