Allow external docker container control
Setting the `SWEETPEA_EXTERNAL_DOCKER_MGMT` environment variable will
cause sweetpea to _not_ perform any docker management. This is useful
in scripts where repeated calls are made the require the container to
be running, and it's faster than starting/stopping the container for
each one.

Also parameterized the timeout for `sharpSAT`. I also had to make some
tweaks to the `sharpSAT` binary to get it to work the way I
wanted. (For example, the `-t` flag is broken in `12.08.1`)