Implement naive solution counting
This starts to introduce a code path where we compute the solution
count for a design, and give that directly to the unigen binary,
rather than relying on another tool (`sharpSAT` or `ApproxMC`) to
approximate the solution count for us.

As indicated in a comment, this mimics the logic found in the
`UniGen.py` script, which is the recommended interface.

At the moment, it naively uses the factorial size of the design, which
is always an upper-bound for our current 'fully-crossed' designs.

It's ok that this isn't perfect yet, since it only matters when
invoking `unigen`, and that doesn't work most of the time yet
anyways.