Add facility for automated end-to-end tests
While working on debugging issues with the `Transition` window, (See
also [this issue][1]) I realized that we're probably going to need to
overhaul how some derived levels are encoded.

In preparation for that, I'd like to have some automated end-to-end
tests that I can run to quickly know if I've broken something that was
working, while we're working on the encoding logic.

[1]: https://github.com/anniecherk/sweetpea-py/issues/5