Refactor internal code structure
It's getting too hard to reason about the different pieces all
together in the same file. I'm going to break things up so that it's
easier to fix the `Transition`/`Window` use cases.

This isn't 100% backwards compatible, as the module structure has
changed, so `import` statements may need to be changed, but the
public classes and function calls stayed the same.