Stop depending on Factor ordering
A number of algorithms were assuming that the design was given with
Factors in a specific order. In particular, they assumed that the all
the factors _without_ complex windows came first, which may not always
be the case.

I considered sorting the design/crossing list before doing anything
else, but it seems wiser to fix the code to not assume.

* Made 'first_variable_for_level' order-agnostic.
* Made sure 'decode_variable' is order-agnostic.
* Make fully-cross constraint exclude derived levels from crossing.
  (I don't think this makes sense to allow)
* Made sure consistency constraint is order-agnostic.
* Made sure nomorethankinarow constraint is order-agnostic.
* Made derivation processor order-agnostic.