Use all non-complex window factors in fully cross counting
While making the TSS acceptance test do more permutations in the design
ordering, another bug was revealed where the counting in the fully cross
constraint was off.

Even though factors with derived levels are _not_ (currently) allowed in
the crossing, they _should_ be accounted for when doing the counting,
as long as they don't have a complex window.

This ensures that the counting accounts for all factors in the grid.