Treat Transition derivation functions properly
For the `WithinTrial` window, we simply pass each level name directly to
the derivation function, as there will always only ever be one value for
each factor that the derived level depends on.

For `Transition`, and more generally `Window` windows, there will be `n`
values for each factor, where `n` is the width of the window. As such,
the intent is that the derivation function will receive a list of values
for each factor.