Add extra test (skipped) to derivation processor
I was starting to work on this, but realized that the code has, up to
this point, assumed that only `WithinTrial` constraints have a width
of 1.

I think the test is valid, in that it can be useful to have windows
with a width of 1, but a stride greater than 1.

I'm going to refactor the code to handle all windows the same, rather
than special casing any of them. `WithinTrial` and `Transition` should
just become syntactic sugar for preconfigured `Window` expressions.