Avoid branching on window type
This refactors all the existing code to _stop_ branching based on the
type of Window, (`WithinTrial`, `Transition`, or `Window`) and instead
only branch on general characteristics, such as width and stride.

This is important because a `Window` declared with width=1 and
stride=1 should behave just like a `WithinTrial` window, not
differently.

Depending on the situation, we branch based on whether or not the
factor is derived, or if the window is 'complex', which means either
the width or the stride is greater than one. (Which subsequently means
that the factor will not have a representation in every trial.)