Keep pointer to a Derivation's associated Factor
The derivation constraint already needed access to the number of levels
in its associated factor in order to correctly compute the encoding
variables. As I was thinking about how to make it account for
strides != 1, I realized it was going to need more information.

This reshapes the derivation constraint to keep a pointer to the entire
associated factor so that all necessary information is available.