Allow derived factors in the crossing
After discussion, we've realized that factors with derived levels _may_
be allowed in the crossing, so long no two factors in the crossing
depend on a common ancestral factor.

This removes the error that was preventing this, and adds an acceptance
test that ensures it works correctly for derived factors when the window
width and stride are 1.

It doesn't work for more complicated windows yet, nor does it do the
ancestor check. That will come later.