Add helpers to block class
This adds a couple more helpers to the block class for obtaining the
variables that correspond to particular factors and levels for a given
trial.

This should make it easier to refactor the `FullyCross` constraint to
deal correctly with derived levels, rather than further complicating the
algebra that exists there already.