Refactor `FullyCross` constraint
This refactors the `FullyCross` constraint to use the new helper methods
from the previous commit, which significantly reduces the complexity of
the constraint implementation.

Next we'll add a test for the `FullyCross` constraint when a
`Transition` is included in the crossing. If I haven't forgotten
anything, it should just work.