Raise error when crossing is invalid
All factors in a crossing must be indepdendent, meaning one cannot be
derived from another, and still be crossed.

This implements a check for that constraint by determining if any factor
in the crossing can be reached from any other factor in the crossing
using the design graph. If such a path exists, that indicates a
dependency, and the crossing must be rejected.