Change graph error to warning
We realized this morning that there are in fact some crossings involving
related factors that are valid (and work), so we shouldn't be
prohibiting all of them.

This does make it easier to create UNSAT designs, so I'm leaving the
error in place as a warning until we can get better debugging help in
place.