Implement cheap constraint desugaring
This isn't perfect, but it's an ok way to do what I need to without
breaking our contract with users.

Constraints have been pretty odd, in that they can accept either a
`Factor`, _or_ an individual factor/level tuple. There has been special
code in the application logic to repeatedly apply the constraint for
each level when the user specifies a factor.

This has worked fine so far.

Now that I'm starting to implement logic to do solution counting, I need
to inspect the constraints. Having to repeat the same special logic to
flatten the factor would be gross.

As a result, I've changed constraints to have a `desugar` function,
which is responsible for doing any real flattening/desugaring that needs
to be done.

It is now the responsibility of the block construction functions to
invoke the `desugar` function, and replace the constraint with anything
it returns.