Require K > 0 on KInARow constraints
This commit does a few things, but the primary motivator was to disallow
k=0 on the `KInARow` constraints. This leaves the `Exclude` constraint
as the _only_ way of prohibiting a particular level, which will simplify
things down the road, especially when they want to exclude derived
levels that are part of the crossing.

As part of this, I realized that the `Exclude` constraint was written
incorrectly. To fix it, I needed to compute the variable list for the
level, which was only in the `_KInARow` base class, so I shifted that
into the `Block`, where anyone could use it if needed.

Lastly, had to fix up the ATs that were using K=0.