Remove placeholder AT
Sebastian pointed out today that a nested derived level could always be
represented just by making the derivation function for a single derived
level a little more complicated. He's right of course, and it simplifies
things quite a bit on the backend, so I think that's the way to go for
now.