Constrain usage of derived levels
This adds an explicit error if you attempt to derive a level 'B' from
another derived level 'A' when 'A' does not apply to every trial. (It
has a window width or stride that is not 1)

If you tried to do this before, it was allowed, but broken. This just
makes the reality clear to the user.

This is needed now that we plan on letting the `Exclude` constraint be
applied to derived levels that may reduce the length of the trial
sequence. When excluded, a derived level with width=stride=1 _does_
reduce the trial sequence length by 1, but derived levels with other
windows may not.