Allow override to relax full crossing constraint
There are some experiments where users want an almost-complete crossing,
meaning they want to start with a full crossing, but then exclude some
number of specific combinations. This hasn't been possible yet, since
the `FullyCrossedBlock` requires a complete crossing.

This commit introduces flag on the `fully_cross_block` constructor
function: `require_complete_crossing`, which can be set to `False` in
order to relax the constraint by reducing the number of trials in the
sequence.

It only reduces the sequence length for `Exclude` constraints that are
applied to factors which apply to each trial.