Improve error reporting for invalid constraints
A user had an error today just because they forgot to include the factor
that they were constraining in the design. Sadly, the error they
received was not helpful at all to them, and they needed help figuring
out the problem.

This change adds additional validation logic so that we identify these
cases and give the user a helpful error message that tells them what to
look for to fix the problem.