Generalize application of `Exclude`
We've been attacking `Exclude` piecemeal, but I think this finishes it
off.

Previously, we were assuming that an exclude constraint would only
remove a single trial from the sequence. In hindsight, that's obviously
not true. Depending on the derivation function used in the exclusion,
there could be an arbitrary number of trials removed.

This corrects the algorithm in `__count_exclusions` to apply the
derivation function of the exclusion to _every_ possible crossing, and
then count the unique number of crossings excluded after checking each
constraint.

There could be a large number of crossings to check, but this should
still run quickly. Any performance impact is a necessary evil, as we
cannot know how many to exclude without checking the derivation function
against each possibility.