Decrement `fresh` before giving it to the backend
The backend assumes that the value of `fresh` represents the last
allocated variable, rather than the next available variable. As a
result, the frontend needs to decrement it before handing it off to
avoid skipping a variable in the sequence, and thus leaving it
unconstrained.

This was already fixed [once][1], but was regressed as part of a later
effort to [reorganize the code][2].

We didn't detect this sooner because it isn't a problem when using
`synthesize_trials_non_uniform`. This is because the non-uniform
generation only negates the support variables in a solution, and adds
those to the CNF. It does _not_ include _all_ variables in a negated
solution. As a result, we can't get the same solution twice.

It does manifest when using tools like sharpSAT, spur, or unigen to
count/sample solutions.

[1]: https://github.com/sweetpea-org/sweetpea-py/commit/9e5f537de6a84d3e67e2c5b8f6a7d59fa5cc277d
[2]: https://github.com/sweetpea-org/sweetpea-py/commit/284bce925744286c2c9b801754da6dc1f76e5c88