Implement rejection sampling for UC sampler
`KInARow` constraints represent roughly the same kind of problem as
enumerating sequences containing `Transition` factors, or derived
factors with other complex windows. The UC sampler can't handle those
yet in sample construction, so this implements a simple rejection
sampling pass to ensure that no constraints are violated. It also
includes metrics in the sampling result to indicate what proportion of
the samples were rejected.

It's not likely that I'll be able to figure out how to enumerate
sequences with constraints or complex windows before my time is up, but
I do hope to at least be able to count them. That way we can at least
provide an estimate of the average rejection count before a valid sample
is constructed.

For example, if we count 10,000 sequences in a design when ignoring
constraints, and then we determine that the constraints will exclude
6,000 of those, then we know that, on average, 1.67 samples will be
rejected before a valid sequence is found.