Fix bugs in `shift_window`
We had a bug in `shift_window` where it wasn't properly accounting for
the number of arguments to the derivation function. It assumed that each
entry in the index list was for a subsequent trial, but this isn't the
case if there were 2 or more arguments to the function.

This repairs `shift_window` by splitting the indices into sublists of
size n, where n is the number of arguments to the derivation function,
and then applying the offset uniformly within each sublist. (Since each
sublist represents the 'next' trial.)