Repair variable list construction
`__build_complex_variable_list` had a bug in that it assumed that there
were only ever two levels for a derived factor with a complex window. As
soon as we tried a design that included a Transition with three levels,
it broke. This repairs the function take the number of levels of the
factor into account.

Resolves https://github.com/sweetpea-org/sweetpea-py/issues/1.