core/cnf: implement CNF.prepend and use it where CNF.append was used
In testing it was discovered that outputs were in the wrong order. After
debugging, it was determined that this was caused by the programmer not
dissociating the original function name ("append") from its
implementation (prepending a new clause to an existing CNF formula).

This commit rectifies the error by implementing a `CNF.prepend` function and
using it in all the places where `CNF.append` was previously used. The original
`CNF.append` implementation is left in place for symmetry.