primitives: remove Generic Factor implementation
Although sensible, the Generic implementation of the base Factor combined with
the desire to use the `Factor` class's `__new__` as a factory for `SimpleFactor`
and `DerivedFactor` caused issues with type-checking. Specifically, variables
instantiated using `Factor` were being required to provide type annotations to
specify the generic type, which is undesirable.

Therefore, the `Generic` superclass was removed from the `Factor`. This requires
overloading `Level`-returning methods in the subclasses such that type checking
works correctly, which is not phenomenal but gets the job done.