fix min_trials to preserve crossing balanace
This commit partly reverts beebe23fc, which changed the crossing
constraint from "less than 2" to "more than 0". Using "more than 0"
was somewhat sensible in the context of that commit, where the
handling of excluded conditions had changed since the original "less
than 2", and so even "equal to 1" without min_trials. However, "more
than 0" is not right for a min_trials value greater than the crossing
size, because it doesn't constrain the second multiple of the crossing
size. For example, 4 possible conditions and a min_trials of 6, it's
possible to get three instances of one condition.

To support min_trials, a better choice is "less than N+1" where N is
min_trials divided by the crossing size rounded up. That preserves
balance overall. It does not, however, constrain the first
crossing-size trials to be balanced, so I think it's still not what we
want. But committing for the record.

This commit also removes the copying of `FullyCross` and
`MultipleCross`, since the latter works for the former.