repair combinatoric sampler with minimim_trials
The handling of a minimum_trials constraint was not right. Although
the intent of the earlier implementation was to pick distinct
sequences for repetitions of a crossing within a generated run, the
correct behavior is to pick sequences independently. Furthermore, the
total number of solutions should cover all possible combinations of
those soutions, not just the component crossing-sized sequences within
runs. Finally, handling a trial count that is not a multiple of the
crossing size requires counting the number of solutions in the case
that fewer than a crossing-size number of tirals will be used (i.e.,
all possible completions up to a crossing-size number would be
considered equivalent).

Meanwhile, the use of the combinatoric sampler had been excessively
constrained. It can deal with derived factors as part of the crossing,
as long as they are within-trial, so that's fixed here, too.