using saturating adder for popcount comparisons
When comparing the popcount of N variables to k, we can use log(k)+1
bits instead of log(N)+1 bits with a saturating add instead of an
exact add. Often, such as for the crossing constraint, k is 1, so we
save bits. For the example experiments, the encoding bcomes 25%
smaller, while CNF-generation time (typically the bottleneck for
non-uniform sampling) is cut in half. The savings are not enough to
make uniform sampling tractable on those examples.