repair sample generation for combinatoric with minimum trials
When a minimum-trials constraint extends the number of trials to be
not a multiple of the crossing size, the uniform combinatoric sampler
was reoprting the right total number of trials, but it was not
correctly varying the non-crossing-sized tail, and so it was reporting
duplicates.

The repair here is in the way the sequence number -> permutation
calculation is performed so that it's compatible with prefixes.