change combinatoric sampling of designs with complex crossed factors
This change reduces the search space for rejection sampling with
complex derived factors. For example, PadmalaPessoa2011 does down from
a seach space of 2^463 to 2^253, while Stroop_simple goes from 2^122
to 2^106, and Gratton goes from 2^63 to 2^32. That's not enough to
make rejection sampling work for exact solutions in the first two of
those examples, but the new search turns out to have a more useful of
fit that allows almost-correct solutions to be found in a practical
amount of time. Instantiate `UniformCombinatoricSamplingStrategy` with
an amount of `acceptable_error` (= missing crossings).

A "complex" factor is one with a window size > 1. The previous
sampling approach was to consider all possible combinations of complex
derived levels with other levels in the crossing. That approach
ensured that the crossing constraint would be fulfilled, but the
chance that derived levels match the derivation rule is miniscule.

The new approach is to select only combinations of non-derived levels
in the crossing, compute the derived levels, and then check whether
the crossing constraint is satisfied. Still, the chances are miniscule
for many examples, but the chance of having only a few combinations in
the crossing are missed turns out to be much higher than the old
method's chance of having only a few derivations miscomputed.