bring back separate API documentation
The SweetPea implementation is not organized in a way that makes it
amenable to automatic documentation extraction. Part of the problem is
core organization, and part of the problem is a bunch of bindings that
we want to keep for backward compatibility. Generating internal
documentation might be useful for some SwetPea maintainers, but that
can be done separately.

Also, clean up factor and level naming and identity. Factor names must
be strings. Level names an be anything, and they don't have to be
distinct.