fix slow handling of excludes
The handling of excludes for the CNF encoding went the long way
around, building up too many possibilities times too many
possibilities, then filtering after that product of those two, where
it turns out that the filtering is uniform over the first dimension of
the product. So, filter diretcly on that first input, instead, and
don't both with an extra layer of `product`.

Also, simplify and repair counting of excluded elements in a crossing.