implement weighted levels
In this implementation weighted levels, a weight affects only crossing
combinations. This choice allows weighting of derived levels, where
duplicating level instead would not make sense. It behaves differently
than might be expected for a factor that is not crossed, however; a
further revision seemed merited.

This commit also fixes problems with the RandomGen sampling strategy
when there are uncrossed factors that are dependencies of crossed
derived factors and either (1) a minimum-trials contract creates extra
trials that are not a multiple of the crossing size; or (2) some
crossing combinations are excluded by contraints or by a derived
level's definition. The implementation previously overcounted some
combinations, which could make sampling non-uniform.