change weighting to affect distribution of levels for uncrossed factors
For non-derived factors, make the effect of weighting the same as
having multiple levels with the same name. When a factor is used only
in crossings, then it is not actually duplicated, since the weight can
be handled directly (without increasing the total number of
solutions). Otherwise, the implementation desugars into multiple
levels.

To map constraints on a single level to constraints on the multiple
replacements of a level, a desugared factor is represented by two
factors: a flattened on and a derived factor that connects a single
level to multipe flattend levels.

With this form of weighting in place, we go back to disallowing
multiple levels with the same name in a factor. Allowing multiple
names creates trouble for constraints that refre to a level by name,
and that problem is magnified by desugaring.